# Faster integer encoding