This PR adds faster versions of the `strconv.Append(U)Int` functions from the stdlib.

```
name           old time/op    new time/op    delta
CodeEncoder-4    2.51ms ± 1%    2.31ms ± 2%   -7.72%  (p=0.000 n=9+9)

name           old speed      new speed      delta
CodeEncoder-4   774MB/s ± 1%   838MB/s ± 2%   +8.37%  (p=0.000 n=9+9)
```

The algorithm is the same — the gains come from:
- inlining the append calls by moving [these](https://github.com/golang/go/blob/ed78c90a784f9703857c3303d871b973ee2f0102/src/strconv/itoa.go#L41-L43) [chunks](https://github.com/golang/go/blob/ed78c90a784f9703857c3303d871b973ee2f0102/src/strconv/itoa.go#L51-L53) of code into the [inner](https://github.com/golang/go/blob/ed78c90a784f9703857c3303d871b973ee2f0102/src/strconv/itoa.go#L44) function which already couldn't be inlined
- replacing the [lookup table of bytes](https://github.com/golang/go/blob/ed78c90a784f9703857c3303d871b973ee2f0102/src/strconv/itoa.go#L68) with a lookup table of `uint16`, and copying two bytes at a time rather than one
- removing unneeded functionality [such as](https://github.com/golang/go/blob/ed78c90a784f9703857c3303d871b973ee2f0102/src/strconv/itoa.go#L90-L92) supporting bases other than 10

```
name                  time/op
AppendIntStd1-4       5.93ns ± 0%
AppendIntStdMinI64-4  33.7ns ± 0%
```
```
name                  time/op
AppendInt1-4          3.47ns ± 1%
AppendIntMinI64-4     22.0ns ± 1%
```

I haven't ported across the `host32bit` [branch](https://github.com/golang/go/blob/ed78c90a784f9703857c3303d871b973ee2f0102/src/strconv/itoa.go#L109) and so the encoding of `(u)int64` values may take a hit on 32-bit systems.