@achille-roussel 2020-09-17
This looks great!

You should contribute some of these back to Go :+1: